[FEA] EFD-Reinf - Ajustes e melhorias diversas
