+++ pysped/efdreinf/leiaute/schema/v1.03.02/XSD Comunicação EFD-REINF v1.3.02 WSDL
