+++ pysped/efdreinf/leiaute/schema/v1.03.02/envioLoteEventos.xsd
